# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a backup management server for Postgres databases.",
        "title": "Backup Manager API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/executions/{id}": {
            "get": {
                "description": "Get full details of a specific backup execution, including logs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "executions"
                ],
                "summary": "Get execution details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Execution ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_repository_db.Execution"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_pkg_apiutil.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_pkg_apiutil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_pkg_apiutil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/executions/{id}/pin": {
            "post": {
                "description": "Prevent a backup execution from being automatically cleaned up by pinning it",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "executions"
                ],
                "summary": "Pin or unpin a backup execution",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Execution ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Pin status (e.g. {'pinned': true})",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_pkg_apiutil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_pkg_apiutil.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/profiles": {
            "get": {
                "description": "Get a list of all configured backup profiles",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profiles"
                ],
                "summary": "List all backup profiles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_repository_db.Profile"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_pkg_apiutil.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new backup profile with the specified configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profiles"
                ],
                "summary": "Create a new backup profile",
                "parameters": [
                    {
                        "description": "Profile configuration",
                        "name": "profile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_repository_db.CreateProfileParams"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_repository_db.Profile"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_pkg_apiutil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_pkg_apiutil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/profiles/{id}": {
            "put": {
                "description": "Update a backup profile by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profiles"
                ],
                "summary": "Update an existing backup profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Profile ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated profile configuration",
                        "name": "profile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_repository_db.UpdateProfileParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_repository_db.Profile"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_pkg_apiutil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_pkg_apiutil.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a backup profile by its ID",
                "tags": [
                    "profiles"
                ],
                "summary": "Delete a backup profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Profile ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_pkg_apiutil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_pkg_apiutil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/profiles/{id}/activate": {
            "post": {
                "description": "Set a profile as the active one for scheduled backups. All other profiles will be deactivated.",
                "tags": [
                    "profiles"
                ],
                "summary": "Activate a backup profile (sets as scheduled)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Profile ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_repository_db.Profile"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_pkg_apiutil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_pkg_apiutil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/profiles/{id}/executions": {
            "get": {
                "description": "Get a history of all backup executions for a profile",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "executions"
                ],
                "summary": "List executions for a specific profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Profile ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_repository_db.Execution"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_pkg_apiutil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_pkg_apiutil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/profiles/{id}/run": {
            "post": {
                "description": "Trigger a backup execution for a specific profile now",
                "tags": [
                    "profiles"
                ],
                "summary": "Run a backup profile immediately",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Profile ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MD2SA_backup-manager_internal_pkg_apiutil.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_MD2SA_backup-manager_internal_pkg_apiutil.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "github_com_MD2SA_backup-manager_internal_repository_db.CreateProfileParams": {
            "type": "object",
            "properties": {
                "compressionLevel": {
                    "type": "integer",
                    "format": "int32"
                },
                "compressionType": {
                    "type": "string"
                },
                "description": {
                    "$ref": "#/definitions/pgtype.Text"
                },
                "enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notificationProviderID": {
                    "type": "string"
                },
                "retentionPolicyID": {
                    "type": "string"
                },
                "schedule": {
                    "type": "string"
                },
                "storageProviderID": {
                    "type": "string"
                }
            }
        },
...
        "github_com_MD2SA_backup-manager_internal_repository_db.Execution": {
            "type": "object",
            "properties": {
                "checksum": {
                    "$ref": "#/definitions/pgtype.Text"
                },
                "createdAt": {
                    "$ref": "#/definitions/pgtype.Timestamptz"
                },
                "duration": {
                    "$ref": "#/definitions/pgtype.Interval"
                },
                "endTime": {
                    "$ref": "#/definitions/pgtype.Timestamptz"
                },
                "errorMessage": {
                    "$ref": "#/definitions/pgtype.Text"
                },
                "id": {
                    "type": "string"
                },
                "isPinned": {
                    "type": "boolean"
                },
                "logs": {
                    "type": "array",
                    "items": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "profileID": {
                    "type": "string"
                },
                "size": {
                    "$ref": "#/definitions/pgtype.Int8"
                },
                "startTime": {
                    "$ref": "#/definitions/pgtype.Timestamptz"
                },
                "status": {
                    "type": "string"
                },
                "storagePath": {
                    "$ref": "#/definitions/pgtype.Text"
                }
            }
        },
...
        "github_com_MD2SA_backup-manager_internal_repository_db.Profile": {
            "type": "object",
            "properties": {
                "compressionLevel": {
                    "type": "integer",
                    "format": "int32"
                },
                "compressionType": {
                    "type": "string"
                },
                "createdAt": {
                    "$ref": "#/definitions/pgtype.Timestamptz"
                },
                "description": {
                    "$ref": "#/definitions/pgtype.Text"
                },
                "enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notificationProviderID": {
                    "type": "string"
                },
                "retentionPolicyID": {
                    "type": "string"
                },
                "schedule": {
                    "type": "string"
                },
                "storageProviderID": {
                    "type": "string"
                },
                "updatedAt": {
                    "$ref": "#/definitions/pgtype.Timestamptz"
                }
            }
        },
...
        "github_com_MD2SA_backup-manager_internal_repository_db.UpdateProfileParams": {
            "type": "object",
            "properties": {
                "compressionLevel": {
                    "type": "integer",
                    "format": "int32"
                },
                "compressionType": {
                    "type": "string"
                },
                "description": {
                    "$ref": "#/definitions/pgtype.Text"
                },
                "enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notificationProviderID": {
                    "type": "string"
                },
                "retentionPolicyID": {
                    "type": "string"
                },
                "schedule": {
                    "type": "string"
                },
                "storageProviderID": {
                    "type": "string"
                }
            }
        },
...
        "pgtype.InfinityModifier": {
            "type": "integer",
            "format": "int32",
            "enum": [
                1,
                0,
                -1
            ],
            "x-enum-varnames": [
                "Infinity",
                "Finite",
                "NegativeInfinity"
            ]
        },
        "pgtype.Int8": {
            "type": "object",
            "properties": {
                "int64": {
                    "type": "integer",
                    "format": "int64"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
        "pgtype.Interval": {
            "type": "object",
            "properties": {
                "days": {
                    "type": "integer",
                    "format": "int32"
                },
                "microseconds": {
                    "type": "integer",
                    "format": "int64"
                },
                "months": {
                    "type": "integer",
                    "format": "int32"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
        "pgtype.Text": {
            "type": "object",
            "properties": {
                "string": {
                    "type": "string"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
        "pgtype.Timestamptz": {
            "type": "object",
            "properties": {
                "infinityModifier": {
                    "$ref": "#/definitions/pgtype.InfinityModifier"
                },
                "time": {
                    "type": "string"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        }
    }
}