# internal/repository/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Execution struct {
	ID           pgtype.UUID
	ProfileID    pgtype.UUID
	Status       string
	StartTime    pgtype.Timestamptz
	EndTime      pgtype.Timestamptz
	Duration     pgtype.Interval
	Size         pgtype.Int8
	Checksum     pgtype.Text
	StoragePath  pgtype.Text
	Logs         []byte
	ErrorMessage pgtype.Text
	IsPinned     bool
	CreatedAt    pgtype.Timestamptz
}

type NotificationProvider struct {
	ID        pgtype.UUID
	Name      string
	Type      string
	Config    []byte
	CreatedAt pgtype.Timestamptz
	UpdatedAt pgtype.Timestamptz
}

type Profile struct {
	ID                     pgtype.UUID
	Name                   string
	Description            pgtype.Text
	Enabled                bool
	Schedule               string
	StorageProviderID      pgtype.UUID
	NotificationProviderID pgtype.UUID
	RetentionPolicyID      pgtype.UUID
	CompressionType        string
	CompressionLevel       int32
	CreatedAt              pgtype.Timestamptz
	UpdatedAt              pgtype.Timestamptz
}

type RetentionPolicy struct {
	ID          pgtype.UUID
	Name        string
	KeepHourly  int32
	KeepDaily   int32
	KeepWeekly  int32
	KeepMonthly int32
	KeepYearly  int32
	YearlyMonth int32
	CreatedAt   pgtype.Timestamptz
	UpdatedAt   pgtype.Timestamptz
}

type StorageProvider struct {
	ID        pgtype.UUID
	Name      string
	Type      string
	Config    []byte
	CreatedAt pgtype.Timestamptz
	UpdatedAt pgtype.Timestamptz
}